refactor: 🏄🏽  move entities to separate module
